elemToManifest moved to Elems
